pyproject.toml
[project]
name = "agent-one-api"
version = "1.0.0"
description = "FastAPI RAG service for Agent One"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic-ai>=0.7.4",
    "asyncpg>=0.30.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
